Refactoring code for system icons in case we need it in different places
