+++ Snoop/Snoop.csproj
@@ -188,6 +188,7 @@
     <Compile Include="Infrastructure\SnoopModes.cs" />
     <Compile Include="Infrastructure\SnoopPartsRegistry.cs" />
     <Compile Include="Infrastructure\ResourceKeyCache.cs" />
+    <Compile Include="Infrastructure\SystemIconHelper.cs" />
     <Compile Include="Infrastructure\TemplateHelper.cs" />
     <Compile Include="Infrastructure\WhenLoaded.cs" />
     <Compile Include="Infrastructure\ZoomerUtilities.cs" />
